+++ spec/PluginClientSpec.php
@@ -88,7 +88,7 @@ class PluginClientSpec extends ObjectBehavior
         $this->shouldThrow('Http\Client\Common\Exception\LoopException')->duringSendRequest($request);
     }
 
-    function it_injects_debug_plugins(HttpClient $httpClient, RequestInterface $request, Plugin $plugin0, Plugin $plugin1, Plugin $debugPlugin)
+    function it_injects_debug_plugins(HttpClient $httpClient, ResponseInterface $response, RequestInterface $request, Plugin $plugin0, Plugin $plugin1, Plugin $debugPlugin)
     {
         $plugin0
             ->handleRequest(
@@ -125,6 +125,7 @@ class PluginClientSpec extends ObjectBehavior
             })
         ;
 
+        $httpClient->sendRequest($request)->willReturn($response);
 
         $this->beConstructedWith($httpClient, [$plugin0, $plugin1], ['debug_plugins'=>[$debugPlugin]]);
         $this->sendRequest($request);